fix: remove unused import, shadowed variable, and unnecessary cast
- Remove unused prefix_sum_scalar_with_carry import in vpa_test
- Remove shadowed data binding in csv_test
- Remove redundant u32 cast on trailing_zeros() in AVX-512 decode

+++ rust/findchars/src/kernel/avx512.rs
@@ -171,7 +171,7 @@ unsafe fn process_chunk(
         let mut bits = nonzero;
         let mut i = 0;
         while bits != 0 {
-            let lane = bits.trailing_zeros() as u32;
+            let lane = bits.trailing_zeros();
             *pos_ptr.add(i) = base_offset as u32 + lane;
             i += 1;
             bits &= bits - 1;
